Authorization Service/requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
pytest>=6.0.0